+++ lib/micha_core.dart
@@ -1,10 +1,13 @@
 library micha_core;
 
 export 'src/async/async_builder.dart';
+export 'src/async/rate_limiter.dart';
+export 'src/async/retried.dart';
 export 'src/async/spinner.dart';
 export 'src/async/wait_for.dart';
 export 'src/extensions/ansi_formatting.dart';
 export 'src/extensions/by_name_or_null.dart';
+export 'src/extensions/clamp.dart';
 export 'src/extensions/get_or_put.dart';
 export 'src/extensions/null_when_empty.dart';
 export 'src/extensions/separated.dart';